# docs/api/openapi.json
{
  "openapi" : "3.1.0",
...
  "components" : {
    "schemas" : {
      "CreateBody" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : [ "string", "null" ]
          },
          "slug" : {
            "type" : [ "string", "null" ]
          }
        }
      },
      "CreateBody1" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : [ "string", "null" ]
          },
          "slug" : {
            "type" : [ "string", "null" ]
          },
          "description" : {
            "type" : [ "string", "null" ]
          },
          "baseLanguageTag" : {
            "type" : [ "string", "null" ]
          }
        }
      },
...
      "UpdateBody1" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : [ "string", "null" ]
          }
        }
      },
      "UpdateBody2" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : [ "string", "null" ]
          },
          "description" : {
            "type" : [ "string", "null" ]
          }
        }
      }
    },
    "securitySchemes" : {
      "SecurityScheme" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT",
        "description" : "Authentication"
      }
    }
  },
  "paths" : {
...
    "/api/v1/organizations" : {
      "get" : {
        "summary" : "List organizations the caller is a member of.",
        "tags" : [ "organizations" ],
        "responses" : {
          "200" : {
            "description" : "Listing returned."
          },
          "401" : {
            "description" : "Not authenticated."
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      },
      "post" : {
        "summary" : "Create a new organization; caller becomes OWNER.",
        "tags" : [ "organizations" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateBody"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Organization created."
          },
          "400" : {
            "description" : "Validation failed."
          },
          "401" : {
            "description" : "Not authenticated."
          },
          "409" : {
            "description" : "Slug already in use."
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      }
    },
    "/api/v1/organizations/{orgSlug}" : {
      "patch" : {
        "summary" : "Rename an organization (name only for v0.1.0).",
        "tags" : [ "organizations" ],
        "parameters" : [ {
          "name" : "orgSlug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateBody1"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Updated."
          },
          "400" : {
            "description" : "Validation failed."
          },
          "401" : {
            "description" : "Not authenticated."
          },
          "404" : {
            "description" : "Not found or caller is not a member."
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      },
      "get" : {
        "summary" : "Fetch one organization the caller belongs to.",
        "tags" : [ "organizations" ],
        "parameters" : [ {
          "name" : "orgSlug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Found."
          },
          "401" : {
            "description" : "Not authenticated."
          },
          "404" : {
            "description" : "Not found or caller is not a member."
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      }
    },
...
    "/api/v1/organizations/{orgSlug}/projects" : {
      "get" : {
        "summary" : "List every project in the organization.",
        "tags" : [ "projects" ],
        "parameters" : [ {
          "name" : "orgSlug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Listing returned."
          },
          "401" : {
            "description" : "Not authenticated."
          },
          "404" : {
            "description" : "Caller is not a member."
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      },
      "post" : {
        "summary" : "Create a project inside the organization.",
        "tags" : [ "projects" ],
        "parameters" : [ {
          "name" : "orgSlug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateBody1"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Project created."
          },
          "400" : {
            "description" : "Validation failed."
          },
          "401" : {
            "description" : "Not authenticated."
          },
          "404" : {
            "description" : "Caller is not a member."
          },
          "409" : {
            "description" : "Slug already in use."
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      }
    },
    "/api/v1/organizations/{orgSlug}/projects/{projectSlug}" : {
      "patch" : {
        "summary" : "Update name or description.",
        "tags" : [ "projects" ],
        "parameters" : [ {
          "name" : "orgSlug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "projectSlug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateBody2"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Updated."
          },
          "400" : {
            "description" : "Validation failed."
          },
          "401" : {
            "description" : "Not authenticated."
          },
          "404" : {
            "description" : "Not found or caller is not a member."
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      },
      "get" : {
        "summary" : "Fetch one project.",
        "tags" : [ "projects" ],
        "parameters" : [ {
          "name" : "orgSlug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "projectSlug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Found."
          },
          "401" : {
            "description" : "Not authenticated."
          },
          "404" : {
            "description" : "Not found or caller is not a member."
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      }
    },
...
  },
  "info" : {
    "title" : "Translately API",
    "version" : "0.0.1-SNAPSHOT",
    "description" : "Open-source self-hosted localization and translation management API.",
    "contact" : {
      "name" : "Pratiyush",
      "url" : "https://github.com/Pratiyush/translately"
    },
    "license" : {
      "name" : "MIT",
      "url" : "https://opensource.org/licenses/MIT"
    }
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ]
}